+++ frontend/src/pages/projects/components/GenericHorizontalBar.tsx
@@ -1,0 +1,68 @@
+import * as React from 'react';
+import {
+  Tabs,
+  Tab,
+  TabTitleIcon,
+  TabTitleText,
+  TabContent,
+  TabContentBody,
+  PageSection,
+  PageSectionVariants,
+} from '@patternfly/react-core';
+import { SVGIconProps } from '@patternfly/react-icons/dist/esm/createIcon';
+
+type GenericHorizontalBarProps = {
+  sections: {
+    title: string;
+    component: React.ReactNode;
+    icon?: React.ReactElement<React.ComponentClass<SVGIconProps>>;
+  }[];
+};
+
+const GenericHorizontalBar: React.FC<GenericHorizontalBarProps> = ({ sections }) => {
+  const [activeTabKey, setActiveTabKey] = React.useState<string | number | undefined>(
+    sections[0].title || undefined,
+  );
+
+  return (
+    <>
+      <PageSection variant={PageSectionVariants.light} type="tabs" isFilled>
+        <Tabs
+          activeKey={activeTabKey}
+          onSelect={(event, tabIndex) => setActiveTabKey(tabIndex)}
+          aria-label="Horizontal bar"
+          mountOnEnter
+        >
+          {sections.map((section) => (
+            <Tab
+              key={section.title}
+              eventKey={section.title}
+              tabContentId={section.title}
+              title={
+                <>
+                  {section.icon && <TabTitleIcon>{section.icon}</TabTitleIcon>}
+                  <TabTitleText>{section.title}</TabTitleText>
+                </>
+              }
+            />
+          ))}
+        </Tabs>
+      </PageSection>
+      <PageSection variant={PageSectionVariants.light} isFilled padding={{ default: 'noPadding' }}>
+        {sections.map((section) => (
+          <TabContent
+            id={section.title}
+            activeKey={activeTabKey}
+            eventKey={section.title}
+            key={section.title}
+            hidden={section.title !== activeTabKey}
+          >
+            <TabContentBody>{section.component}</TabContentBody>
+          </TabContent>
+        ))}
+      </PageSection>
+    </>
+  );
+};
+
+export default GenericHorizontalBar;
