+++ frontend/src/pages/projects/projectSharing/utils.ts
@@ -1,0 +1,5 @@
+import { DashboardConfig } from '~/types';
+import { featureFlagEnabled } from '~/utilities/utils';
+
+export const isProjectSharingEnabled = (dashboardConfig: DashboardConfig) =>
+  featureFlagEnabled(dashboardConfig.spec.dashboardConfig.disableProjectSharing);
