+++ frontend/src/pages/projects/screens/detail/ProjectDetails.tsx
@@ -1,89 +1,25 @@
 import * as React from 'react';
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  Divider,
-  PageSection,
-  Stack,
-  StackItem,
-} from '@patternfly/react-core';
+import { Breadcrumb, BreadcrumbItem } from '@patternfly/react-core';
+import { CubeIcon, UsersIcon } from '@patternfly/react-icons';
 import { Link } from 'react-router-dom';
 import ApplicationsPage from '~/pages/ApplicationsPage';
-import GenericSidebar from '~/components/GenericSidebar';
-import { useAppContext } from '~/app/AppContext';
-import ServingRuntimeList from '~/pages/modelServing/screens/projects/ServingRuntimeList';
 import { ProjectDetailsContext } from '~/pages/projects/ProjectDetailsContext';
 import { getProjectDescription, getProjectDisplayName } from '~/pages/projects/utils';
-import { featureFlagEnabled } from '~/utilities/utils';
-import PipelinesSection from '~/pages/projects/screens/detail/pipelines/PipelinesSection';
-import NotebooksList from './notebooks/NotebookList';
-import { ProjectSectionID } from './types';
-import StorageList from './storage/StorageList';
-import { ProjectSectionTitles } from './const';
-import DataConnectionsList from './data-connections/DataConnectionsList';
+import GenericHorizontalBar from '~/pages/projects/components/GenericHorizontalBar';
+import ProjectSharing from '~/pages/projects/projectSharing/ProjectSharing';
+import { useAppContext } from '~/app/AppContext';
+import { isProjectSharingEnabled } from '~/pages/projects/projectSharing/utils';
 import useCheckLogoutParams from './useCheckLogoutParams';
-
-type SectionType = {
-  id: ProjectSectionID;
-  component: React.ReactNode;
-  isEmpty: boolean;
-};
+import ProjectDetailsComponents from './ProjectDetailsComponents';
 
 const ProjectDetails: React.FC = () => {
-  const {
-    currentProject,
-    notebooks: { data: notebookStates, loaded: notebookStatesLoaded },
-    pvcs: { data: pvcs, loaded: pvcsLoaded },
-    dataConnections: { data: connections, loaded: connectionsLoaded },
-    servingRuntimes: { data: modelServers, loaded: modelServersLoaded },
-  } = React.useContext(ProjectDetailsContext);
+  const { dashboardConfig } = useAppContext();
+  const { currentProject } = React.useContext(ProjectDetailsContext);
   const displayName = getProjectDisplayName(currentProject);
   const description = getProjectDescription(currentProject);
-  const { dashboardConfig } = useAppContext();
-  const modelServingEnabled = featureFlagEnabled(
-    dashboardConfig.spec.dashboardConfig.disableModelServing,
-  );
-  const pipelinesEnabled = featureFlagEnabled(
-    dashboardConfig.spec.dashboardConfig.disablePipelines,
-  );
-  useCheckLogoutParams();
+  const projectSharingEnabled = isProjectSharingEnabled(dashboardConfig);
 
-  const scrollableSelectorID = 'project-details-list';
-  const sections: SectionType[] = [
-    {
-      id: ProjectSectionID.WORKBENCHES,
-      component: <NotebooksList />,
-      isEmpty: notebookStatesLoaded && notebookStates.length === 0,
-    },
-    {
-      id: ProjectSectionID.CLUSTER_STORAGES,
-      component: <StorageList />,
-      isEmpty: pvcsLoaded && pvcs.length === 0,
-    },
-    {
-      id: ProjectSectionID.DATA_CONNECTIONS,
-      component: <DataConnectionsList />,
-      isEmpty: connectionsLoaded && connections.length === 0,
-    },
-    ...(pipelinesEnabled
-      ? [
-          {
-            id: ProjectSectionID.PIPELINES,
-            component: <PipelinesSection />,
-            isEmpty: pipelinesEnabled,
-          },
-        ]
-      : []),
-    ...(modelServingEnabled
-      ? [
-          {
-            id: ProjectSectionID.MODEL_SERVER,
-            component: <ServingRuntimeList />,
-            isEmpty: modelServersLoaded && modelServers.length === 0,
-          },
-        ]
-      : []),
-  ];
+  useCheckLogoutParams();
 
   return (
     <ApplicationsPage
@@ -98,37 +34,16 @@ const ProjectDetails: React.FC = () => {
       loaded
       empty={false}
     >
-      <PageSection
-        id={scrollableSelectorID}
-        hasOverflowScroll
-        isFilled
-        aria-label="project-details-page-section"
-        variant="light"
-      >
-        <GenericSidebar
-          sections={sections.map(({ id }) => id)}
-          titles={ProjectSectionTitles}
-          scrollableSelector={`#${scrollableSelectorID}`}
-          maxWidth={175}
-        >
-          <Stack hasGutter>
-            {sections.map(({ id, component, isEmpty }, index) => (
-              <React.Fragment key={id}>
-                <StackItem
-                  id={id}
-                  aria-label={ProjectSectionTitles[id]}
-                  data-testid="details-page-section"
-                >
-                  {component}
-                </StackItem>
-                {index !== sections.length - 1 && isEmpty && (
-                  <Divider data-testid="details-page-section-divider" />
-                )}
-              </React.Fragment>
-            ))}
-          </Stack>
-        </GenericSidebar>
-      </PageSection>
+      {projectSharingEnabled ? (
+        <GenericHorizontalBar
+          sections={[
+            { title: 'Components', component: <ProjectDetailsComponents />, icon: <CubeIcon /> },
+            { title: 'Permissions', component: <ProjectSharing />, icon: <UsersIcon /> },
+          ]}
+        />
+      ) : (
+        <ProjectDetailsComponents />
+      )}
     </ApplicationsPage>
   );
 };
